Multiple changes
Add exit function to close component and prop validate it. Also remove current method of closing which was by unmount at node with reactDOM. Change function to get user details from variable to a state being handled with useEffect. Add await to username check which fixes the bug of it not working as intended. Also change logic for determining start source of the avatara and background images.